requirements.txt
flask
torch
torchvision
pandas
open-clip-torch
scikit-learn